Fix extracting of files in big offsets (were int instead of long)
git-svn-id: https://gtav-modding.googlecode.com/svn/trunk@11 d3e79847-3c27-d15e-6f65-b16d5ed7a0e8

+++ LibertyV/RPF7/Entries/Entry.cs
@@ -51,7 +51,7 @@ namespace LibertyV.RPF7.Entries
             BitsStream stream = new BitsStream(new MemoryStream(data));
 
             bool isResource = stream.ReadBool();
-            int offset = (int)stream.ReadBits(23);
+            long offset = (long)stream.ReadBits(23);
             int compressedSize = (int)stream.ReadBits(24);
             int filenameOffset = (int)stream.ReadBits(16);
